made loging recet entity and repo

+++ loginservice/src/main/java/org/ulearn/login/loginservice/repository/LoginRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.ulearn.login.loginservice.entity.LoginEntity;
 
 
-public interface LoginRepository extends JpaRepository<LoginEntity, Integer> {
+public interface LoginRepository extends JpaRepository<LoginEntity, Long> {
 	Optional<LoginEntity> findByUserName(String userName);
 }
